watcher/requirements.txt
kubernetes>=26.0.0
GitPython>=3.1.30
PyYAML>=6.0
